+++ test/integration/p11_algorithms_integration_test.go
@@ -25,7 +25,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/ThalesGroup/k8s-kms-plugin/pkg/providers"
+	"github.com/eclipse-keysealer/k8s-kms-plugin/pkg/providers"
 
 	k8skmsv2 "k8s.io/kms/apis/v2"
 )